# connectorapi-go-pico/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API Gateway for ConnectorAPI.",
        "title": "Connector API Gateway",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "SYE",
            "url": "https://aeon.co.th",
            "email": "sye@aeon.co.th"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "10.254.97.103:8082",
    "basePath": "/",
    "paths": {
...
        "/Api/Agreement/UpdateStatus": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agreement"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "API key",
                        "name": "Api-Key",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "DeviceOS",
                        "name": "Api-DeviceOS",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Channel",
                        "name": "Api-Channel",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "RequestID",
                        "name": "Api-RequestID",
                        "in": "header"
                    },
                    {
                        "description": "BodyRequest",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateStatusResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.UpdateStatusRequest": {
            "type": "object",
            "required": [
                "AEONID",
                "Agreement",
                "Status"
            ],
            "properties": {
                "AEONID": {
                    "type": "string",
                    "maxLength": 20
                },
                "Agreement": {
                    "type": "string",
                    "maxLength": 12
                },
                "Status": {
                    "type": "string",
                    "maxLength": 1
                }
            }
        },
        "domain.UpdateStatusResponse": {
            "type": "object",
            "properties": {
                "AEONID": {
                    "type": "string",
                    "maxLength": 20
                },
                "Agreement": {
                    "type": "string",
                    "maxLength": 12
                }
            }
        }
    }
}